EA-001 - created global exception handler and user response with calculations (#1)
* EA-001 - created global exception handler and user response with calculations

* Delete .idea/sonarlint/issuestore/c/6/c6a1f6a717d32dfcd31c0000d1a6122aa76c348a

* Delete .idea/sonarlint/issuestore/index.pb

* Delete .idea/sonarlint/securityhotspotstore/index.pb

* Delete .idea/workspace.xml

---------

+++ user/user-application/src/main/java/pl/java/user/application/in/response/UserResponse.java
@@ -10,6 +10,6 @@ public record UserResponse(
         String type,
         String avatarUrl,
         String createdAt,
-        int calculations
+        double calculations
 ) {
 }
